Crate aws_sdk_glacier

Amazon S3 Glacier (Glacier) is a storage solution for "cold data."

Glacier is an extremely low-cost storage service that provides secure, durable, and easy-to-use storage for data backup and archival. With Glacier, customers can store their data cost effectively for months, years, or decades. Glacier also enables customers to offload the administrative burdens of operating and scaling storage to AWS, so they don't have to worry about capacity planning, hardware provisioning, data replication, hardware failure and recovery, or time-consuming hardware migrations.

Glacier is a great storage choice when low storage cost is paramount and your data is rarely retrieved. If your application requires fast or frequent access to your data, consider using Amazon S3. For more information, see Amazon Simple Storage Service (Amazon S3).

You can store any kind of data in any format. There is no maximum limit on the total amount of data you can store in Glacier.

If you are a first-time user of Glacier, we recommend that you begin by reading the following sections in the Amazon S3 Glacier Developer Guide:

What is Amazon S3 Glacier - This section of the Developer Guide describes the underlying data model, the operations it supports, and the AWS SDKs that you can use to interact with the service.

Getting Started with Amazon S3 Glacier - The Getting Started section walks you through the process of creating a vault, uploading archives, creating jobs to download archives, retrieving the job output, and deleting archives.

Re-exports

pub use config::Config;

pub use client::Client;

Modules

client

Client and fluent builders for calling the service.

config

Configuration for the service.

error

Errors that can occur when calling the service.

input

Input structures for operations.

model

Data structures used by operation inputs/outputs.

operation

All operations that this crate can perform.

output

Output structures for operations.

Structs

Blob

Binary Blob Type

ByteStream

Stream of binary data

Credentials

AWS SDK Credentials

Endpoint

API Endpoint

Region

The region to send requests to.

RetryConfig

Retry configuration for requests.

Enums

Error

All possible error types for this service.

SdkError

Failed SDK Result

Statics

PKG_VERSION

Crate version number.