Module aws_sdk_glacier::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AbortMultipartUpload

Fluent builder constructing a request to AbortMultipartUpload.

AbortVaultLock

Fluent builder constructing a request to AbortVaultLock.

AddTagsToVault

Fluent builder constructing a request to AddTagsToVault.

CompleteMultipartUpload

Fluent builder constructing a request to CompleteMultipartUpload.

CompleteVaultLock

Fluent builder constructing a request to CompleteVaultLock.

CreateVault

Fluent builder constructing a request to CreateVault.

DeleteArchive

Fluent builder constructing a request to DeleteArchive.

DeleteVault

Fluent builder constructing a request to DeleteVault.

DeleteVaultAccessPolicy

Fluent builder constructing a request to DeleteVaultAccessPolicy.

DeleteVaultNotifications

Fluent builder constructing a request to DeleteVaultNotifications.

DescribeJob

Fluent builder constructing a request to DescribeJob.

DescribeVault

Fluent builder constructing a request to DescribeVault.

GetDataRetrievalPolicy

Fluent builder constructing a request to GetDataRetrievalPolicy.

GetJobOutput

Fluent builder constructing a request to GetJobOutput.

GetVaultAccessPolicy

Fluent builder constructing a request to GetVaultAccessPolicy.

GetVaultLock

Fluent builder constructing a request to GetVaultLock.

GetVaultNotifications

Fluent builder constructing a request to GetVaultNotifications.

InitiateJob

Fluent builder constructing a request to InitiateJob.

InitiateMultipartUpload

Fluent builder constructing a request to InitiateMultipartUpload.

InitiateVaultLock

Fluent builder constructing a request to InitiateVaultLock.

ListJobs

Fluent builder constructing a request to ListJobs.

ListMultipartUploads

Fluent builder constructing a request to ListMultipartUploads.

ListParts

Fluent builder constructing a request to ListParts.

ListProvisionedCapacity

Fluent builder constructing a request to ListProvisionedCapacity.

ListTagsForVault

Fluent builder constructing a request to ListTagsForVault.

ListVaults

Fluent builder constructing a request to ListVaults.

PurchaseProvisionedCapacity

Fluent builder constructing a request to PurchaseProvisionedCapacity.

RemoveTagsFromVault

Fluent builder constructing a request to RemoveTagsFromVault.

SetDataRetrievalPolicy

Fluent builder constructing a request to SetDataRetrievalPolicy.

SetVaultAccessPolicy

Fluent builder constructing a request to SetVaultAccessPolicy.

SetVaultNotifications

Fluent builder constructing a request to SetVaultNotifications.

UploadArchive

Fluent builder constructing a request to UploadArchive.

UploadMultipartPart

Fluent builder constructing a request to UploadMultipartPart.