Struct aws_sdk_glacier::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for Glacier.

This client allows ergonomic access to a Glacier-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn abort_multipart_upload(&self) -> AbortMultipartUpload<C, M, R>

Constructs a fluent builder for the AbortMultipartUpload operation.

See AbortMultipartUpload for more information about the operation and its arguments.

pub fn abort_vault_lock(&self) -> AbortVaultLock<C, M, R>

Constructs a fluent builder for the AbortVaultLock operation.

See AbortVaultLock for more information about the operation and its arguments.

pub fn add_tags_to_vault(&self) -> AddTagsToVault<C, M, R>

Constructs a fluent builder for the AddTagsToVault operation.

See AddTagsToVault for more information about the operation and its arguments.

pub fn complete_multipart_upload(&self) -> CompleteMultipartUpload<C, M, R>

Constructs a fluent builder for the CompleteMultipartUpload operation.

See CompleteMultipartUpload for more information about the operation and its arguments.

pub fn complete_vault_lock(&self) -> CompleteVaultLock<C, M, R>

Constructs a fluent builder for the CompleteVaultLock operation.

See CompleteVaultLock for more information about the operation and its arguments.

pub fn create_vault(&self) -> CreateVault<C, M, R>

Constructs a fluent builder for the CreateVault operation.

See CreateVault for more information about the operation and its arguments.

pub fn delete_archive(&self) -> DeleteArchive<C, M, R>

Constructs a fluent builder for the DeleteArchive operation.

See DeleteArchive for more information about the operation and its arguments.

pub fn delete_vault(&self) -> DeleteVault<C, M, R>

Constructs a fluent builder for the DeleteVault operation.

See DeleteVault for more information about the operation and its arguments.

pub fn delete_vault_access_policy(&self) -> DeleteVaultAccessPolicy<C, M, R>

Constructs a fluent builder for the DeleteVaultAccessPolicy operation.

See DeleteVaultAccessPolicy for more information about the operation and its arguments.

pub fn delete_vault_notifications(&self) -> DeleteVaultNotifications<C, M, R>

Constructs a fluent builder for the DeleteVaultNotifications operation.

See DeleteVaultNotifications for more information about the operation and its arguments.

pub fn describe_job(&self) -> DescribeJob<C, M, R>

Constructs a fluent builder for the DescribeJob operation.

See DescribeJob for more information about the operation and its arguments.

pub fn describe_vault(&self) -> DescribeVault<C, M, R>

Constructs a fluent builder for the DescribeVault operation.

See DescribeVault for more information about the operation and its arguments.

pub fn get_data_retrieval_policy(&self) -> GetDataRetrievalPolicy<C, M, R>

Constructs a fluent builder for the GetDataRetrievalPolicy operation.

See GetDataRetrievalPolicy for more information about the operation and its arguments.

pub fn get_job_output(&self) -> GetJobOutput<C, M, R>

Constructs a fluent builder for the GetJobOutput operation.

See GetJobOutput for more information about the operation and its arguments.

pub fn get_vault_access_policy(&self) -> GetVaultAccessPolicy<C, M, R>

Constructs a fluent builder for the GetVaultAccessPolicy operation.

See GetVaultAccessPolicy for more information about the operation and its arguments.

pub fn get_vault_lock(&self) -> GetVaultLock<C, M, R>

Constructs a fluent builder for the GetVaultLock operation.

See GetVaultLock for more information about the operation and its arguments.

pub fn get_vault_notifications(&self) -> GetVaultNotifications<C, M, R>

Constructs a fluent builder for the GetVaultNotifications operation.

See GetVaultNotifications for more information about the operation and its arguments.

pub fn initiate_job(&self) -> InitiateJob<C, M, R>

Constructs a fluent builder for the InitiateJob operation.

See InitiateJob for more information about the operation and its arguments.

pub fn initiate_multipart_upload(&self) -> InitiateMultipartUpload<C, M, R>

Constructs a fluent builder for the InitiateMultipartUpload operation.

See InitiateMultipartUpload for more information about the operation and its arguments.

pub fn initiate_vault_lock(&self) -> InitiateVaultLock<C, M, R>

Constructs a fluent builder for the InitiateVaultLock operation.

See InitiateVaultLock for more information about the operation and its arguments.

pub fn list_jobs(&self) -> ListJobs<C, M, R>

Constructs a fluent builder for the ListJobs operation.

See ListJobs for more information about the operation and its arguments.

pub fn list_multipart_uploads(&self) -> ListMultipartUploads<C, M, R>

Constructs a fluent builder for the ListMultipartUploads operation.

See ListMultipartUploads for more information about the operation and its arguments.

pub fn list_parts(&self) -> ListParts<C, M, R>

Constructs a fluent builder for the ListParts operation.

See ListParts for more information about the operation and its arguments.

pub fn list_provisioned_capacity(&self) -> ListProvisionedCapacity<C, M, R>

Constructs a fluent builder for the ListProvisionedCapacity operation.

See ListProvisionedCapacity for more information about the operation and its arguments.

pub fn list_tags_for_vault(&self) -> ListTagsForVault<C, M, R>

Constructs a fluent builder for the ListTagsForVault operation.

See ListTagsForVault for more information about the operation and its arguments.

pub fn list_vaults(&self) -> ListVaults<C, M, R>

Constructs a fluent builder for the ListVaults operation.

See ListVaults for more information about the operation and its arguments.

pub fn purchase_provisioned_capacity(
&self
) -> PurchaseProvisionedCapacity<C, M, R>

Constructs a fluent builder for the PurchaseProvisionedCapacity operation.

See PurchaseProvisionedCapacity for more information about the operation and its arguments.

pub fn remove_tags_from_vault(&self) -> RemoveTagsFromVault<C, M, R>

Constructs a fluent builder for the RemoveTagsFromVault operation.

See RemoveTagsFromVault for more information about the operation and its arguments.

pub fn set_data_retrieval_policy(&self) -> SetDataRetrievalPolicy<C, M, R>

Constructs a fluent builder for the SetDataRetrievalPolicy operation.

See SetDataRetrievalPolicy for more information about the operation and its arguments.

pub fn set_vault_access_policy(&self) -> SetVaultAccessPolicy<C, M, R>

Constructs a fluent builder for the SetVaultAccessPolicy operation.

See SetVaultAccessPolicy for more information about the operation and its arguments.

pub fn set_vault_notifications(&self) -> SetVaultNotifications<C, M, R>

Constructs a fluent builder for the SetVaultNotifications operation.

See SetVaultNotifications for more information about the operation and its arguments.

pub fn upload_archive(&self) -> UploadArchive<C, M, R>

Constructs a fluent builder for the UploadArchive operation.

See UploadArchive for more information about the operation and its arguments.

pub fn upload_multipart_part(&self) -> UploadMultipartPart<C, M, R>

Constructs a fluent builder for the UploadMultipartPart operation.

See UploadMultipartPart for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.