Module aws_sdk_glacier::error

Errors that can occur when calling the service.

Modules

insufficient_capacity_exception

See InsufficientCapacityException

invalid_parameter_value_exception

See InvalidParameterValueException

limit_exceeded_exception

See LimitExceededException

missing_parameter_value_exception

See MissingParameterValueException

policy_enforced_exception

See PolicyEnforcedException

request_timeout_exception

See RequestTimeoutException

resource_not_found_exception

See ResourceNotFoundException

service_unavailable_exception

See ServiceUnavailableException

Structs

AbortMultipartUploadError

Error type for the AbortMultipartUpload operation.

AbortVaultLockError

Error type for the AbortVaultLock operation.

AddTagsToVaultError

Error type for the AddTagsToVault operation.

CompleteMultipartUploadError

Error type for the CompleteMultipartUpload operation.

CompleteVaultLockError

Error type for the CompleteVaultLock operation.

CreateVaultError

Error type for the CreateVault operation.

DeleteArchiveError

Error type for the DeleteArchive operation.

DeleteVaultAccessPolicyError

Error type for the DeleteVaultAccessPolicy operation.

DeleteVaultError

Error type for the DeleteVault operation.

DeleteVaultNotificationsError

Error type for the DeleteVaultNotifications operation.

DescribeJobError

Error type for the DescribeJob operation.

DescribeVaultError

Error type for the DescribeVault operation.

GetDataRetrievalPolicyError

Error type for the GetDataRetrievalPolicy operation.

GetJobOutputError

Error type for the GetJobOutput operation.

GetVaultAccessPolicyError

Error type for the GetVaultAccessPolicy operation.

GetVaultLockError

Error type for the GetVaultLock operation.

GetVaultNotificationsError

Error type for the GetVaultNotifications operation.

InitiateJobError

Error type for the InitiateJob operation.

InitiateMultipartUploadError

Error type for the InitiateMultipartUpload operation.

InitiateVaultLockError

Error type for the InitiateVaultLock operation.

InsufficientCapacityException

Returned if there is insufficient capacity to process this expedited request. This

InvalidParameterValueException

Returned if a parameter of the request is incorrectly specified.

LimitExceededException

Returned if the request results in a vault or account limit being exceeded.

ListJobsError

Error type for the ListJobs operation.

ListMultipartUploadsError

Error type for the ListMultipartUploads operation.

ListPartsError

Error type for the ListParts operation.

ListProvisionedCapacityError

Error type for the ListProvisionedCapacity operation.

ListTagsForVaultError

Error type for the ListTagsForVault operation.

ListVaultsError

Error type for the ListVaults operation.

MissingParameterValueException

Returned if a required header or parameter is missing from the request.

PolicyEnforcedException

Returned if a retrieval job would exceed the current data policy's retrieval rate

PurchaseProvisionedCapacityError

Error type for the PurchaseProvisionedCapacity operation.

RemoveTagsFromVaultError

Error type for the RemoveTagsFromVault operation.

RequestTimeoutException

Returned if, when uploading an archive, Amazon S3 Glacier times out while receiving the

ResourceNotFoundException

Returned if the specified resource (such as a vault, upload ID, or job ID) doesn't

ServiceUnavailableException

Returned if the service cannot complete the request.

SetDataRetrievalPolicyError

Error type for the SetDataRetrievalPolicy operation.

SetVaultAccessPolicyError

Error type for the SetVaultAccessPolicy operation.

SetVaultNotificationsError

Error type for the SetVaultNotifications operation.

UploadArchiveError

Error type for the UploadArchive operation.

UploadMultipartPartError

Error type for the UploadMultipartPart operation.

Enums

AbortMultipartUploadErrorKind

Types of errors that can occur for the AbortMultipartUpload operation.

AbortVaultLockErrorKind

Types of errors that can occur for the AbortVaultLock operation.

AddTagsToVaultErrorKind

Types of errors that can occur for the AddTagsToVault operation.

CompleteMultipartUploadErrorKind

Types of errors that can occur for the CompleteMultipartUpload operation.

CompleteVaultLockErrorKind

Types of errors that can occur for the CompleteVaultLock operation.

CreateVaultErrorKind

Types of errors that can occur for the CreateVault operation.

DeleteArchiveErrorKind

Types of errors that can occur for the DeleteArchive operation.

DeleteVaultAccessPolicyErrorKind

Types of errors that can occur for the DeleteVaultAccessPolicy operation.

DeleteVaultErrorKind

Types of errors that can occur for the DeleteVault operation.

DeleteVaultNotificationsErrorKind

Types of errors that can occur for the DeleteVaultNotifications operation.

DescribeJobErrorKind

Types of errors that can occur for the DescribeJob operation.

DescribeVaultErrorKind

Types of errors that can occur for the DescribeVault operation.

GetDataRetrievalPolicyErrorKind

Types of errors that can occur for the GetDataRetrievalPolicy operation.

GetJobOutputErrorKind

Types of errors that can occur for the GetJobOutput operation.

GetVaultAccessPolicyErrorKind

Types of errors that can occur for the GetVaultAccessPolicy operation.

GetVaultLockErrorKind

Types of errors that can occur for the GetVaultLock operation.

GetVaultNotificationsErrorKind

Types of errors that can occur for the GetVaultNotifications operation.

InitiateJobErrorKind

Types of errors that can occur for the InitiateJob operation.

InitiateMultipartUploadErrorKind

Types of errors that can occur for the InitiateMultipartUpload operation.

InitiateVaultLockErrorKind

Types of errors that can occur for the InitiateVaultLock operation.

ListJobsErrorKind

Types of errors that can occur for the ListJobs operation.

ListMultipartUploadsErrorKind

Types of errors that can occur for the ListMultipartUploads operation.

ListPartsErrorKind

Types of errors that can occur for the ListParts operation.

ListProvisionedCapacityErrorKind

Types of errors that can occur for the ListProvisionedCapacity operation.

ListTagsForVaultErrorKind

Types of errors that can occur for the ListTagsForVault operation.

ListVaultsErrorKind

Types of errors that can occur for the ListVaults operation.

PurchaseProvisionedCapacityErrorKind

Types of errors that can occur for the PurchaseProvisionedCapacity operation.

RemoveTagsFromVaultErrorKind

Types of errors that can occur for the RemoveTagsFromVault operation.

SetDataRetrievalPolicyErrorKind

Types of errors that can occur for the SetDataRetrievalPolicy operation.

SetVaultAccessPolicyErrorKind

Types of errors that can occur for the SetVaultAccessPolicy operation.

SetVaultNotificationsErrorKind

Types of errors that can occur for the SetVaultNotifications operation.

UploadArchiveErrorKind

Types of errors that can occur for the UploadArchive operation.

UploadMultipartPartErrorKind

Types of errors that can occur for the UploadMultipartPart operation.