Module aws_sdk_glacier::input

Input structures for operations.

Modules

abort_multipart_upload_input

See AbortMultipartUploadInput

abort_vault_lock_input

See AbortVaultLockInput

add_tags_to_vault_input

See AddTagsToVaultInput

complete_multipart_upload_input

See CompleteMultipartUploadInput

complete_vault_lock_input

See CompleteVaultLockInput

create_vault_input

See CreateVaultInput

delete_archive_input

See DeleteArchiveInput

delete_vault_access_policy_input

See DeleteVaultAccessPolicyInput

delete_vault_input

See DeleteVaultInput

delete_vault_notifications_input

See DeleteVaultNotificationsInput

describe_job_input

See DescribeJobInput

describe_vault_input

See DescribeVaultInput

get_data_retrieval_policy_input

See GetDataRetrievalPolicyInput

get_job_output_input

See GetJobOutputInput

get_vault_access_policy_input

See GetVaultAccessPolicyInput

get_vault_lock_input

See GetVaultLockInput

get_vault_notifications_input

See GetVaultNotificationsInput

initiate_job_input

See InitiateJobInput

initiate_multipart_upload_input

See InitiateMultipartUploadInput

initiate_vault_lock_input

See InitiateVaultLockInput

list_jobs_input

See ListJobsInput

list_multipart_uploads_input

See ListMultipartUploadsInput

list_parts_input

See ListPartsInput

list_provisioned_capacity_input

See ListProvisionedCapacityInput

list_tags_for_vault_input

See ListTagsForVaultInput

list_vaults_input

See ListVaultsInput

purchase_provisioned_capacity_input

See PurchaseProvisionedCapacityInput

remove_tags_from_vault_input

See RemoveTagsFromVaultInput

set_data_retrieval_policy_input

See SetDataRetrievalPolicyInput

set_vault_access_policy_input

See SetVaultAccessPolicyInput

set_vault_notifications_input

See SetVaultNotificationsInput

upload_archive_input

See UploadArchiveInput

upload_multipart_part_input

See UploadMultipartPartInput

Structs

AbortMultipartUploadInput

Provides options to abort a multipart upload identified by the upload ID.

AbortVaultLockInput

The input values for AbortVaultLock.

AddTagsToVaultInput

The input values for AddTagsToVault.

CompleteMultipartUploadInput

Provides options to complete a multipart upload operation. This informs Amazon

CompleteVaultLockInput

The input values for CompleteVaultLock.

CreateVaultInput

Provides options to create a vault.

DeleteArchiveInput

Provides options for deleting an archive from an Amazon S3 Glacier vault.

DeleteVaultAccessPolicyInput

DeleteVaultAccessPolicy input.

DeleteVaultInput

Provides options for deleting a vault from Amazon S3 Glacier.

DeleteVaultNotificationsInput

Provides options for deleting a vault notification configuration from an Amazon

DescribeJobInput

Provides options for retrieving a job description.

DescribeVaultInput

Provides options for retrieving metadata for a specific vault in Amazon

GetDataRetrievalPolicyInput

Input for GetDataRetrievalPolicy.

GetJobOutputInput

Provides options for downloading output of an Amazon S3 Glacier job.

GetVaultAccessPolicyInput

Input for GetVaultAccessPolicy.

GetVaultLockInput

The input values for GetVaultLock.

GetVaultNotificationsInput

Provides options for retrieving the notification configuration set on an Amazon

InitiateJobInput

Provides options for initiating an Amazon S3 Glacier job.

InitiateMultipartUploadInput

Provides options for initiating a multipart upload to an Amazon S3 Glacier

InitiateVaultLockInput

The input values for InitiateVaultLock.

ListJobsInput

Provides options for retrieving a job list for an Amazon S3 Glacier vault.

ListMultipartUploadsInput

Provides options for retrieving list of in-progress multipart uploads for an Amazon

ListPartsInput

Provides options for retrieving a list of parts of an archive that have been uploaded

ListProvisionedCapacityInput

ListTagsForVaultInput

The input value for ListTagsForVaultInput.

ListVaultsInput

Provides options to retrieve the vault list owned by the calling user's account. The

PurchaseProvisionedCapacityInput

RemoveTagsFromVaultInput

The input value for RemoveTagsFromVaultInput.

SetDataRetrievalPolicyInput

SetDataRetrievalPolicy input.

SetVaultAccessPolicyInput

SetVaultAccessPolicy input.

SetVaultNotificationsInput

Provides options to configure notifications that will be sent when specific events

UploadArchiveInput

Provides options to add an archive to a vault.

UploadMultipartPartInput

Provides options to upload a part of an archive in a multipart upload