Struct aws_sdk_glacier::input::UploadMultipartPartInput

#[non_exhaustive]
pub struct UploadMultipartPartInput {
    pub account_id: Option<String>,
    pub vault_name: Option<String>,
    pub upload_id: Option<String>,
    pub checksum: Option<String>,
    pub range: Option<String>,
    pub body: ByteStream,
}

Provides options to upload a part of an archive in a multipart upload operation.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

account_id: Option<String>

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vault_name: Option<String>

The name of the vault.

upload_id: Option<String>

The upload ID of the multipart upload.

checksum: Option<String>

The SHA256 tree hash of the data being uploaded.

range: Option<String>

Identifies the range of bytes in the assembled archive that will be uploaded in this part. Amazon S3 Glacier uses this information to assemble the archive in the proper sequence. The format of this header follows RFC 2616. An example header is Content-Range:bytes 0-4194303/*.

body: ByteStream

The data to upload.

Implementations

impl UploadMultipartPartInput

pub async fn make_operation(
self,
_config: &Config
) -> Result<Operation<UploadMultipartPart, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<UploadMultipartPart>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture UploadMultipartPartInput

impl UploadMultipartPartInput

pub fn account_id(&self) -> Option<&str>

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

pub fn vault_name(&self) -> Option<&str>

The name of the vault.

pub fn upload_id(&self) -> Option<&str>

The upload ID of the multipart upload.

pub fn checksum(&self) -> Option<&str>

The SHA256 tree hash of the data being uploaded.

pub fn range(&self) -> Option<&str>

Identifies the range of bytes in the assembled archive that will be uploaded in this part. Amazon S3 Glacier uses this information to assemble the archive in the proper sequence. The format of this header follows RFC 2616. An example header is Content-Range:bytes 0-4194303/*.

pub fn body(&self) -> &ByteStream

The data to upload.