Module aws_sdk_glacier::model

Data structures used by operation inputs/outputs.

Modules

csv_input

See CsvInput

csv_output

See CsvOutput

data_retrieval_policy

See DataRetrievalPolicy

data_retrieval_rule

See DataRetrievalRule

describe_vault_output

See DescribeVaultOutput

encryption

See Encryption

glacier_job_description

See GlacierJobDescription

grant

See Grant

grantee

See Grantee

input_serialization

See InputSerialization

inventory_retrieval_job_description

See InventoryRetrievalJobDescription

inventory_retrieval_job_input

See InventoryRetrievalJobInput

job_parameters

See JobParameters

output_location

See OutputLocation

output_serialization

See OutputSerialization

part_list_element

See PartListElement

provisioned_capacity_description

See ProvisionedCapacityDescription

s3_location

See S3Location

select_parameters

See SelectParameters

upload_list_element

See UploadListElement

vault_access_policy

See VaultAccessPolicy

vault_lock_policy

See VaultLockPolicy

vault_notification_config

See VaultNotificationConfig

Structs

CsvInput

Contains information about the comma-separated value (CSV) file to select from.

CsvOutput

Contains information about the comma-separated value (CSV) file that the job results

DataRetrievalPolicy

Data retrieval policy.

DataRetrievalRule

Data retrieval policy rule.

DescribeVaultOutput

Contains the Amazon S3 Glacier response to your request.

Encryption

Contains information about the encryption used to store the job results in Amazon S3.

GlacierJobDescription

Contains the description of an Amazon S3 Glacier job.

Grant

Contains information about a grant.

Grantee

Contains information about the grantee.

InputSerialization

Describes how the archive is serialized.

InventoryRetrievalJobDescription

Describes the options for a range inventory retrieval job.

InventoryRetrievalJobInput

Provides options for specifying a range inventory retrieval job.

JobParameters

Provides options for defining a job.

OutputLocation

Contains information about the location where the select job results are stored.

OutputSerialization

Describes how the select output is serialized.

PartListElement

A list of the part sizes of the multipart upload.

ProvisionedCapacityDescription

The definition for a provisioned capacity unit.

S3Location

Contains information about the location in Amazon S3 where the select job results are stored.

SelectParameters

Contains information about the parameters used for a select.

UploadListElement

A list of in-progress multipart uploads for a vault.

VaultAccessPolicy

Contains the vault access policy.

VaultLockPolicy

Contains the vault lock policy.

VaultNotificationConfig

Represents a vault's notification configuration.

Enums

ActionCode

CannedAcl

EncryptionType

ExpressionType

FileHeaderInfo

Permission

QuoteFields

StatusCode

StorageClass

Type