Struct aws_sdk_glacier::output::DescribeJobOutput

#[non_exhaustive]
pub struct DescribeJobOutput {

    pub job_id: Option<String>,
    pub job_description: Option<String>,
    pub action: Option<ActionCode>,
    pub archive_id: Option<String>,
    pub vault_arn: Option<String>,
    pub creation_date: Option<String>,
    pub completed: bool,
    pub status_code: Option<StatusCode>,
    pub status_message: Option<String>,
    pub archive_size_in_bytes: Option<i64>,
    pub inventory_size_in_bytes: Option<i64>,
    pub sns_topic: Option<String>,
    pub completion_date: Option<String>,
    pub sha256_tree_hash: Option<String>,
    pub archive_sha256_tree_hash: Option<String>,
    pub retrieval_byte_range: Option<String>,
    pub tier: Option<String>,
    pub inventory_retrieval_parameters: Option<InventoryRetrievalJobDescription>,
    pub job_output_path: Option<String>,
    pub select_parameters: Option<SelectParameters>,
    pub output_location: Option<OutputLocation>,

}

Contains the description of an Amazon S3 Glacier job.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

job_id: Option<String>

An opaque string that identifies an Amazon S3 Glacier job.

job_description: Option<String>

The job description provided when initiating the job.

action: Option<ActionCode>

The job type. This value is either ArchiveRetrieval, InventoryRetrieval, or Select.

archive_id: Option<String>

The archive ID requested for a select job or archive retrieval. Otherwise, this field is null.

vault_arn: Option<String>

The Amazon Resource Name (ARN) of the vault from which an archive retrieval was requested.

creation_date: Option<String>

The UTC date when the job was created. This value is a string representation of ISO 8601 date format, for example "2012-03-20T17:03:43.221Z".

completed: bool

The job status. When a job is completed, you get the job's output using Get Job Output (GET output).

status_code: Option<StatusCode>

The status code can be InProgress, Succeeded, or Failed, and indicates the status of the job.

status_message: Option<String>

A friendly message that describes the job status.

archive_size_in_bytes: Option<i64>

For an archive retrieval job, this value is the size in bytes of the archive being requested for download. For an inventory retrieval or select job, this value is null.

inventory_size_in_bytes: Option<i64>

For an inventory retrieval job, this value is the size in bytes of the inventory requested for download. For an archive retrieval or select job, this value is null.

sns_topic: Option<String>

An Amazon SNS topic that receives notification.

completion_date: Option<String>

The UTC time that the job request completed. While the job is in progress, the value is null.

sha256_tree_hash: Option<String>

For an archive retrieval job, this value is the checksum of the archive. Otherwise, this value is null.

The SHA256 tree hash value for the requested range of an archive. If the InitiateJob request for an archive specified a tree-hash aligned range, then this field returns a value.

If the whole archive is retrieved, this value is the same as the ArchiveSHA256TreeHash value.

This field is null for the following:

Archive retrieval jobs that specify a range that is not tree-hash aligned

Archival jobs that specify a range that is equal to the whole archive, when the job status is InProgress

Inventory jobs

Select jobs

archive_sha256_tree_hash: Option<String>

The SHA256 tree hash of the entire archive for an archive retrieval. For inventory retrieval or select jobs, this field is null.

retrieval_byte_range: Option<String>

The retrieved byte range for archive retrieval jobs in the form StartByteValue-EndByteValue. If no range was specified in the archive retrieval, then the whole archive is retrieved. In this case, StartByteValue equals 0 and EndByteValue equals the size of the archive minus 1. For inventory retrieval or select jobs, this field is null.

tier: Option<String>

The tier to use for a select or an archive retrieval. Valid values are Expedited, Standard, or Bulk. Standard is the default.

inventory_retrieval_parameters: Option<InventoryRetrievalJobDescription>

Parameters used for range inventory retrieval.

job_output_path: Option<String>

Contains the job output location.

select_parameters: Option<SelectParameters>

Contains the parameters used for a select.

output_location: Option<OutputLocation>

Contains the location where the data from the select job is stored.

Implementations

impl DescribeJobOutput

pub fn job_id(&self) -> Option<&str>

An opaque string that identifies an Amazon S3 Glacier job.

pub fn job_description(&self) -> Option<&str>

The job description provided when initiating the job.

pub fn action(&self) -> Option<&ActionCode>

The job type. This value is either ArchiveRetrieval, InventoryRetrieval, or Select.

pub fn archive_id(&self) -> Option<&str>

The archive ID requested for a select job or archive retrieval. Otherwise, this field is null.

pub fn vault_arn(&self) -> Option<&str>

The Amazon Resource Name (ARN) of the vault from which an archive retrieval was requested.

pub fn creation_date(&self) -> Option<&str>

The UTC date when the job was created. This value is a string representation of ISO 8601 date format, for example "2012-03-20T17:03:43.221Z".

pub fn completed(&self) -> bool

The job status. When a job is completed, you get the job's output using Get Job Output (GET output).

pub fn status_code(&self) -> Option<&StatusCode>

The status code can be InProgress, Succeeded, or Failed, and indicates the status of the job.

pub fn status_message(&self) -> Option<&str>

A friendly message that describes the job status.

pub fn archive_size_in_bytes(&self) -> Option<i64>

For an archive retrieval job, this value is the size in bytes of the archive being requested for download. For an inventory retrieval or select job, this value is null.

pub fn inventory_size_in_bytes(&self) -> Option<i64>

For an inventory retrieval job, this value is the size in bytes of the inventory requested for download. For an archive retrieval or select job, this value is null.

pub fn sns_topic(&self) -> Option<&str>

An Amazon SNS topic that receives notification.

pub fn completion_date(&self) -> Option<&str>

The UTC time that the job request completed. While the job is in progress, the value is null.

pub fn sha256_tree_hash(&self) -> Option<&str>

For an archive retrieval job, this value is the checksum of the archive. Otherwise, this value is null.

The SHA256 tree hash value for the requested range of an archive. If the InitiateJob request for an archive specified a tree-hash aligned range, then this field returns a value.

If the whole archive is retrieved, this value is the same as the ArchiveSHA256TreeHash value.

This field is null for the following:

Archive retrieval jobs that specify a range that is not tree-hash aligned

Archival jobs that specify a range that is equal to the whole archive, when the job status is InProgress

Inventory jobs

Select jobs

pub fn archive_sha256_tree_hash(&self) -> Option<&str>

The SHA256 tree hash of the entire archive for an archive retrieval. For inventory retrieval or select jobs, this field is null.

pub fn retrieval_byte_range(&self) -> Option<&str>

The retrieved byte range for archive retrieval jobs in the form StartByteValue-EndByteValue. If no range was specified in the archive retrieval, then the whole archive is retrieved. In this case, StartByteValue equals 0 and EndByteValue equals the size of the archive minus 1. For inventory retrieval or select jobs, this field is null.

pub fn tier(&self) -> Option<&str>

The tier to use for a select or an archive retrieval. Valid values are Expedited, Standard, or Bulk. Standard is the default.

pub fn inventory_retrieval_parameters(
&self
) -> Option<&InventoryRetrievalJobDescription>

Parameters used for range inventory retrieval.

pub fn job_output_path(&self) -> Option<&str>

Contains the job output location.

pub fn select_parameters(&self) -> Option<&SelectParameters>

Contains the parameters used for a select.

pub fn output_location(&self) -> Option<&OutputLocation>

Contains the location where the data from the select job is stored.

impl DescribeJobOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DescribeJobOutput